+++ packages/import/src/importers/linearCsv/LinearCsvImporter.ts
@@ -69,7 +69,7 @@ export class LinearCsvImporter implements Importer {
         continue;
       }
 
-      const tags = row.Labels.split(", ");
+      const tags = row.Labels ? row.Labels.split(", ") : [];
       const labels = tags.filter(tag => !!tag);
 
       importData.issues.push({