+++ packages/sdk/src/_generated_sdk.ts
@@ -224,6 +224,7 @@ export class AgentActivity extends Request {
     this.createdAt = parseDate(data.createdAt) ?? new Date();
     this.ephemeral = data.ephemeral;
     this.id = data.id;
+    this.signalMetadata = parseJson(data.signalMetadata) ?? undefined;
     this.sourceMetadata = parseJson(data.sourceMetadata) ?? undefined;
     this.updatedAt = parseDate(data.updatedAt) ?? new Date();
     this.signal = data.signal ?? undefined;
@@ -241,6 +242,8 @@ export class AgentActivity extends Request {
   public ephemeral: boolean;
   /** The unique identifier of the entity. */
   public id: string;
+  /** Metadata about this agent activity's signal. */
+  public signalMetadata?: Record<string, unknown>;
   /** Metadata about the external source that created this agent activity. */
   public sourceMetadata?: Record<string, unknown>;
   /**
@@ -459,6 +462,8 @@ export class AgentActivityWebhookPayload {
     this.createdAt = data.createdAt;
     this.id = data.id;
     this.signal = data.signal ?? undefined;
+    this.signalMetadata = data.signalMetadata ?? undefined;
+    this.sourceCommentId = data.sourceCommentId ?? undefined;
     this.updatedAt = data.updatedAt;
     this.userId = data.userId ?? undefined;
   }
@@ -475,6 +480,10 @@ export class AgentActivityWebhookPayload {
   public id: string;
   /** An optional modifier that provides additional instructions on how the activity should be interpreted. */
   public signal?: string;
+  /** Metadata about this agent activity's signal. */
+  public signalMetadata?: L.Scalars["JSONObject"];
+  /** The ID of the comment this activity is linked to. */
+  public sourceCommentId?: string;
   /** The time at which the entity was updated. */
   public updatedAt: string;
   /** The ID of the user who created this agent activity. */
@@ -1362,6 +1371,7 @@ export class AuditEntryWebhookPayload {
 export class AuthIdentityProvider extends Request {
   public constructor(request: LinearRequest, data: L.AuthIdentityProviderFragment) {
     super(request);
+    this.createdAt = parseDate(data.createdAt) ?? new Date();
     this.defaultMigrated = data.defaultMigrated;
     this.id = data.id;
     this.issuerEntityId = data.issuerEntityId ?? undefined;
@@ -1374,6 +1384,8 @@ export class AuthIdentityProvider extends Request {
     this.ssoSigningCert = data.ssoSigningCert ?? undefined;
   }
 
+  /** The time at which the entity was created. */
+  public createdAt: Date;
   /** Whether the identity provider is the default identity provider migrated from organization level settings. */
   public defaultMigrated: boolean;
   /** The unique identifier of the entity. */
@@ -1405,6 +1417,7 @@ export class AuthOrganization extends Request {
   public constructor(request: LinearRequest, data: L.AuthOrganizationFragment) {
     super(request);
     this.allowedAuthServices = data.allowedAuthServices;
+    this.createdAt = parseDate(data.createdAt) ?? new Date();
     this.deletionRequestedAt = parseDate(data.deletionRequestedAt) ?? undefined;
     this.enabled = data.enabled;
     this.id = data.id;
@@ -1422,6 +1435,8 @@ export class AuthOrganization extends Request {
 
   /** Allowed authentication providers, empty array means all are allowed */
   public allowedAuthServices: string[];
+  /** The time at which the entity was created. */
+  public createdAt: Date;
   /** The time at which deletion of the organization was requested. */
   public deletionRequestedAt?: Date;
   /** Whether the organization is enabled. Used as a superuser tool to lock down the org. */
@@ -1502,6 +1517,7 @@ export class AuthUser extends Request {
     super(request);
     this.active = data.active;
     this.avatarUrl = data.avatarUrl ?? undefined;
+    this.createdAt = parseDate(data.createdAt) ?? new Date();
     this.displayName = data.displayName;
     this.email = data.email;
     this.id = data.id;
@@ -1515,6 +1531,8 @@ export class AuthUser extends Request {
   public active: boolean;
   /** An URL to the user's avatar image. */
   public avatarUrl?: string;
+  /** The time at which the entity was created. */
+  public createdAt: Date;
   /** The user's display (nick) name. Unique within each organization. */
   public displayName: string;
   /** The user's email address. */
@@ -1565,7 +1583,7 @@ export class AuthenticationSessionResponse extends Request {
   public client?: string;
   /** Country codes of all seen locations. */
   public countryCodes: string[];
-  /** Date when the session was created. */
+  /** The time at which the entity was created. */
   public createdAt: Date;
   public id: string;
   /** IP address. */
@@ -4903,6 +4921,7 @@ export class EmailIntakeAddress extends Request {
     this.issueCompletedAutoReply = data.issueCompletedAutoReply ?? undefined;
     this.issueCompletedAutoReplyEnabled = data.issueCompletedAutoReplyEnabled;
     this.issueCreatedAutoReply = data.issueCreatedAutoReply ?? undefined;
+    this.issueCreatedAutoReplyEnabled = data.issueCreatedAutoReplyEnabled;
     this.repliesEnabled = data.repliesEnabled;
     this.senderName = data.senderName ?? undefined;
     this.updatedAt = parseDate(data.updatedAt) ?? new Date();
@@ -4940,6 +4959,8 @@ export class EmailIntakeAddress extends Request {
   public issueCompletedAutoReplyEnabled: boolean;
   /** The auto-reply message for issue created. If not set, the default reply will be used. */
   public issueCreatedAutoReply?: string;
+  /** Whether the auto-reply for issue created is enabled. */
+  public issueCreatedAutoReplyEnabled: boolean;
   /** Whether email replies are enabled. */
   public repliesEnabled: boolean;
   /** The name to be used for outgoing emails. */
@@ -5886,6 +5907,21 @@ export class FetchDataPayload extends Request {
   /** The GraphQL query used to fetch the data. */
   public query?: string;
   /** Whether the fetch operation was successful. */
+  public success: boolean;
+}
+/**
+ * FileUploadDeletePayload model
+ *
+ * @param request - function to call the graphql client
+ * @param data - L.FileUploadDeletePayloadFragment response data
+ */
+export class FileUploadDeletePayload extends Request {
+  public constructor(request: LinearRequest, data: L.FileUploadDeletePayloadFragment) {
+    super(request);
+    this.success = data.success;
+  }
+
+  /** Whether the operation was successful. */
   public success: boolean;
 }
 /**
@@ -10231,6 +10267,7 @@ export class IssueSuggestion extends Request {
     super(request);
     this.archivedAt = parseDate(data.archivedAt) ?? undefined;
     this.createdAt = parseDate(data.createdAt) ?? new Date();
+    this.dismissalReason = data.dismissalReason ?? undefined;
     this.id = data.id;
     this.issueId = data.issueId;
     this.stateChangedAt = parseDate(data.stateChangedAt) ?? new Date();
@@ -10253,6 +10290,7 @@ export class IssueSuggestion extends Request {
   public archivedAt?: Date;
   /** The time at which the entity was created. */
   public createdAt: Date;
+  public dismissalReason?: string;
   /** The unique identifier of the entity. */
   public id: string;
   public issueId: string;
@@ -11648,6 +11686,7 @@ export class Organization extends Request {
     this.gitBranchFormat = data.gitBranchFormat ?? undefined;
     this.gitLinkbackMessagesEnabled = data.gitLinkbackMessagesEnabled;
     this.gitPublicLinkbackMessagesEnabled = data.gitPublicLinkbackMessagesEnabled;
+    this.hipaaComplianceEnabled = data.hipaaComplianceEnabled;
     this.id = data.id;
     this.initiativeUpdateReminderFrequencyInWeeks = data.initiativeUpdateReminderFrequencyInWeeks ?? undefined;
     this.initiativeUpdateRemindersHour = data.initiativeUpdateRemindersHour;
@@ -11709,6 +11748,8 @@ export class Organization extends Request {
   public gitLinkbackMessagesEnabled: boolean;
   /** Whether the Git integration linkback messages should be sent to public repositories. */
   public gitPublicLinkbackMessagesEnabled: boolean;
+  /** Whether HIPAA compliance is enabled for organization. */
+  public hipaaComplianceEnabled: boolean;
   /** The unique identifier of the entity. */
   public id: string;
   /** The n-weekly frequency at which to prompt for initiative updates. When not set, reminders are off. */
@@ -15347,6 +15388,82 @@ export class RoadmapToProjectPayload extends Request {
   /** The ID of roadmaptoproject that was created or updated. */
   public get roadmapToProjectId(): string | undefined {
     return this._roadmapToProject?.id;
+  }
+}
+/**
+ * Payload returned by semantic search.
+ *
+ * @param request - function to call the graphql client
+ * @param data - L.SemanticSearchPayloadFragment response data
+ */
+export class SemanticSearchPayload extends Request {
+  public constructor(request: LinearRequest, data: L.SemanticSearchPayloadFragment) {
+    super(request);
+    this.enabled = data.enabled;
+    this.results = data.results.map(node => new SemanticSearchResult(request, node));
+  }
+
+  /** Whether the semantic search is enabled. */
+  public enabled: boolean;
+  public results: SemanticSearchResult[];
+}
+/**
+ * A semantic search result reference.
+ *
+ * @param request - function to call the graphql client
+ * @param data - L.SemanticSearchResultFragment response data
+ */
+export class SemanticSearchResult extends Request {
+  private _document?: L.SemanticSearchResultFragment["document"];
+  private _initiative?: L.SemanticSearchResultFragment["initiative"];
+  private _issue?: L.SemanticSearchResultFragment["issue"];
+  private _project?: L.SemanticSearchResultFragment["project"];
+
+  public constructor(request: LinearRequest, data: L.SemanticSearchResultFragment) {
+    super(request);
+    this.id = data.id;
+    this.type = data.type;
+    this._document = data.document ?? undefined;
+    this._initiative = data.initiative ?? undefined;
+    this._issue = data.issue ?? undefined;
+    this._project = data.project ?? undefined;
+  }
+
+  /** The unique identifier of the entity. */
+  public id: string;
+  /** The type of the semantic search result. */
+  public type: L.SemanticSearchResultType;
+  /** The document related to the semantic search result. */
+  public get document(): LinearFetch<Document> | undefined {
+    return this._document?.id ? new DocumentQuery(this._request).fetch(this._document?.id) : undefined;
+  }
+  /** The ID of document related to the semantic search result. */
+  public get documentId(): string | undefined {
+    return this._document?.id;
+  }
+  /** The initiative related to the semantic search result. */
+  public get initiative(): LinearFetch<Initiative> | undefined {
+    return this._initiative?.id ? new InitiativeQuery(this._request).fetch(this._initiative?.id) : undefined;
+  }
+  /** The ID of initiative related to the semantic search result. */
+  public get initiativeId(): string | undefined {
+    return this._initiative?.id;
+  }
+  /** The issue related to the semantic search result. */
+  public get issue(): LinearFetch<Issue> | undefined {
+    return this._issue?.id ? new IssueQuery(this._request).fetch(this._issue?.id) : undefined;
+  }
+  /** The ID of issue related to the semantic search result. */
+  public get issueId(): string | undefined {
+    return this._issue?.id;
+  }
+  /** The project related to the semantic search result. */
+  public get project(): LinearFetch<Project> | undefined {
+    return this._project?.id ? new ProjectQuery(this._request).fetch(this._project?.id) : undefined;
+  }
+  /** The ID of project related to the semantic search result. */
+  public get projectId(): string | undefined {
+    return this._project?.id;
   }
 }
 /**
@@ -21245,6 +21362,40 @@ export class SearchProjectsQuery extends Request {
 }
 
 /**
+ * A fetchable SemanticSearch Query
+ *
+ * @param request - function to call the graphql client
+ */
+export class SemanticSearchQuery extends Request {
+  public constructor(request: LinearRequest) {
+    super(request);
+  }
+
+  /**
+   * Call the SemanticSearch query and return a SemanticSearchPayload
+   *
+   * @param query - required query to pass to semanticSearch
+   * @param variables - variables without 'query' to pass into the SemanticSearchQuery
+   * @returns parsed response from SemanticSearchQuery
+   */
+  public async fetch(
+    query: string,
+    variables?: Omit<L.SemanticSearchQueryVariables, "query">
+  ): LinearFetch<SemanticSearchPayload> {
+    const response = await this._request<L.SemanticSearchQuery, L.SemanticSearchQueryVariables>(
+      L.SemanticSearchDocument,
+      {
+        query,
+        ...variables,
+      }
+    );
+    const data = response.semanticSearch;
+
+    return new SemanticSearchPayload(this._request, data);
+  }
+}
+
+/**
  * A fetchable SsoUrlFromEmail Query
  *
  * @param request - function to call the graphql client
@@ -26303,6 +26454,35 @@ export class DeleteIssueMutation extends Request {
 }
 
 /**
+ * A fetchable IssueExternalSyncDisable Mutation
+ *
+ * @param request - function to call the graphql client
+ */
+export class IssueExternalSyncDisableMutation extends Request {
+  public constructor(request: LinearRequest) {
+    super(request);
+  }
+
+  /**
+   * Call the IssueExternalSyncDisable mutation and return a IssuePayload
+   *
+   * @param attachmentId - required attachmentId to pass to issueExternalSyncDisable
+   * @returns parsed response from IssueExternalSyncDisableMutation
+   */
+  public async fetch(attachmentId: string): LinearFetch<IssuePayload> {
+    const response = await this._request<
+      L.IssueExternalSyncDisableMutation,
+      L.IssueExternalSyncDisableMutationVariables
+    >(L.IssueExternalSyncDisableDocument, {
+      attachmentId,
+    });
+    const data = response.issueExternalSyncDisable;
+
+    return new IssuePayload(this._request, data);
+  }
+}
+
+/**
  * A fetchable IssueImportCreateAsana Mutation
  *
  * @param request - function to call the graphql client
@@ -29883,6 +30063,35 @@ export class SuspendUserMutation extends Request {
       id,
     });
     const data = response.userSuspend;
+
+    return new UserAdminPayload(this._request, data);
+  }
+}
+
+/**
+ * A fetchable UserUnlinkFromIdentityProvider Mutation
+ *
+ * @param request - function to call the graphql client
+ */
+export class UserUnlinkFromIdentityProviderMutation extends Request {
+  public constructor(request: LinearRequest) {
+    super(request);
+  }
+
+  /**
+   * Call the UserUnlinkFromIdentityProvider mutation and return a UserAdminPayload
+   *
+   * @param id - required id to pass to userUnlinkFromIdentityProvider
+   * @returns parsed response from UserUnlinkFromIdentityProviderMutation
+   */
+  public async fetch(id: string): LinearFetch<UserAdminPayload> {
+    const response = await this._request<
+      L.UserUnlinkFromIdentityProviderMutation,
+      L.UserUnlinkFromIdentityProviderMutationVariables
+    >(L.UserUnlinkFromIdentityProviderDocument, {
+      id,
+    });
+    const data = response.userUnlinkFromIdentityProvider;
 
     return new UserAdminPayload(this._request, data);
   }
@@ -37554,6 +37763,19 @@ export class LinearSdk extends Request {
     return new SearchProjectsQuery(this._request).fetch(term, variables);
   }
   /**
+   * Search for various resources using natural language.
+   *
+   * @param query - required query to pass to semanticSearch
+   * @param variables - variables without 'query' to pass into the SemanticSearchQuery
+   * @returns SemanticSearchPayload
+   */
+  public semanticSearch(
+    query: string,
+    variables?: Omit<L.SemanticSearchQueryVariables, "query">
+  ): LinearFetch<SemanticSearchPayload> {
+    return new SemanticSearchQuery(this._request).fetch(query, variables);
+  }
+  /**
    * Fetch SSO login URL for the email provided.
    *
    * @param email - required email to pass to ssoUrlFromEmail
@@ -39308,6 +39530,15 @@ export class LinearSdk extends Request {
     return new DeleteIssueMutation(this._request).fetch(id, variables);
   }
   /**
+   * Disables external sync on an issue.
+   *
+   * @param attachmentId - required attachmentId to pass to issueExternalSyncDisable
+   * @returns IssuePayload
+   */
+  public issueExternalSyncDisable(attachmentId: string): LinearFetch<IssuePayload> {
+    return new IssueExternalSyncDisableMutation(this._request).fetch(attachmentId);
+  }
+  /**
    * Kicks off an Asana import job.
    *
    * @param asanaTeamName - required asanaTeamName to pass to issueImportCreateAsana
@@ -40535,6 +40766,15 @@ export class LinearSdk extends Request {
     return new SuspendUserMutation(this._request).fetch(id);
   }
   /**
+   * Unlinks a guest user from their identity provider. Can only be called by an admin when SCIM is enabled.
+   *
+   * @param id - required id to pass to userUnlinkFromIdentityProvider
+   * @returns UserAdminPayload
+   */
+  public userUnlinkFromIdentityProvider(id: string): LinearFetch<UserAdminPayload> {
+    return new UserUnlinkFromIdentityProviderMutation(this._request).fetch(id);
+  }
+  /**
    * Un-suspends a user. Can only be called by an admin.
    *
    * @param id - required id to pass to unsuspendUser
@@ -40638,3 +40878,63 @@ export class LinearSdk extends Request {
     return new UpdateWorkflowStateMutation(this._request).fetch(id, input);
   }
 }
+export {
+  AgentActivitySignal,
+  AgentActivityType,
+  AgentSessionStatus,
+  AgentSessionType,
+  AuthenticationSessionType,
+  ContextViewType,
+  CustomerStatusType,
+  CyclePeriod,
+  DateResolutionType,
+  Day,
+  EmailIntakeAddressType,
+  ExternalSyncService,
+  FacetPageSource,
+  FeedSummarySchedule,
+  FrequencyResolutionType,
+  GitAutomationStates,
+  GitLinkKind,
+  GithubOrgType,
+  InitiativeStatus,
+  InitiativeTab,
+  InitiativeUpdateHealthType,
+  IntegrationService,
+  IssueRelationType,
+  IssueSuggestionState,
+  IssueSuggestionType,
+  NotificationCategory,
+  NotificationChannel,
+  OAuthClientApprovalStatus,
+  OrganizationDomainAuthType,
+  OrganizationInviteStatus,
+  OtherNotificationType,
+  PaginationNulls,
+  PaginationOrderBy,
+  PaginationSortOrder,
+  PostType,
+  ProductIntelligenceScope,
+  ProjectMilestoneStatus,
+  ProjectStatusType,
+  ProjectTab,
+  ProjectUpdateHealthType,
+  ProjectUpdateReminderFrequency,
+  PullRequestMergeMethod,
+  PullRequestReviewTool,
+  PullRequestStatus,
+  PushSubscriptionType,
+  ReleaseChannel,
+  SLADayCountType,
+  SemanticSearchResultType,
+  SendStrategy,
+  SlaStatus,
+  SlackChannelType,
+  TriageResponsibilityAction,
+  UserContextViewType,
+  UserFlagType,
+  UserFlagUpdateOperation,
+  UserRoleType,
+  ViewPreferencesType,
+  ViewType,
+} from "./_generated_documents";