+++ packages/codegen-test/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## 2.1.1
+
+### Patch Changes
+
+- 7ae1128: chore(deps): update dependency patch versions
+
+## 2.1.0
+
+### Minor Changes
+
+- feea0fb: Upgrade graphql-codegen from v1 to v2 internally, which produces slight changes in resultant schema fragments and documents.
+
+## 2.0.46
+
+### Patch Changes
+
+- 340065a: chore(deps): update dependency patch versions
+
 ## 2.0.45
 
 ### Patch Changes